refactor: move participant saving into quiz
