+++ src/utils.py
@@ -14,31 +14,3 @@ def safe_quit(win):
 def check_for_quit(win):
     if config.QUIT_KEY in event.getKeys():
         safe_quit(win)
-
-def save_participant_data(participant_info, order, results):
-    participant_id = datetime.now().strftime("%Y%m%d_%H%M%S")
-    
-    file_exists = os.path.isfile(config.PARTICIPANT_DATA)
-    
-    data_row = {
-        'participant_id': participant_id,
-        'timestamp': datetime.now().strftime("%Y-%m-%d"),
-        'age': participant_info.get('age', ''),
-        'gender': participant_info.get('gender', ''),
-        'condition_order': order,
-        'score_black': results.get('score_black', ''),
-        'score_video': results.get('score_video', '')
-    }
-    
-    with open(config.PARTICIPANT_DATA, 'a', newline='', encoding='utf-8') as f:
-        fieldnames = ['participant_id', 'timestamp', 'age', 'gender', 
-                     'condition_order', 'score_black', 'score_video']
-        writer = csv.DictWriter(f, fieldnames=fieldnames)
-        
-        if not file_exists:
-            writer.writeheader()
-        
-        writer.writerow(data_row)
-    
-    print(f"Data saved for participant {participant_id}")
-    return participant_id
